range data wrapper shifted to miluv
