+++ src/ekf_velocity.py
@@ -14,7 +14,6 @@ from pyuwbcalib.utils import (
 )
 from utils.measurement import (
     Measurement,
-    RangeData,
 )
 from utils.misc import (
     GaussianResult,
@@ -35,6 +34,7 @@ from utils.models import (
     AltitudeById,
 )
 from miluv.data import DataLoader
+from miluv.wrappers import RangeData
 from src.filters import ExtendedKalmanFilter
 import time
 from tqdm import tqdm
